+++ packages/view-main/src/modules/app/init.ts
@@ -1,5 +1,6 @@
 // import { createUnsubscriptionSet } from '@/shared'
 import { setTitle } from '@/shared'
+import { handleConfigChange, handleRelease, initWindowInfo } from '@/shared/browser/widnow.svelte'
 import { keyboardEvent } from '../hotkey/keyboard'
 import { lyricEvent } from '../lyric/store/event'
 import { playerEvent } from '../player/store/event'
@@ -7,12 +8,15 @@ import { playerState } from '../player/store/state'
 import { settingEvent } from '../setting/store/event'
 import { settingState } from '../setting/store/state'
 import { onConnected, onRelease } from './shared'
-import { getMachineId, sendInitedEvent, setMachineId, setWorkerInitPromise } from './store/action'
+import { getMachineId, sendInitedEvent, setFullScreen, setMachineId, setWorkerInitPromise } from './store/action'
 import { appEvent } from './store/event'
 
 const init = async () => {
   const machineId = await getMachineId()
   setMachineId(machineId)
+  if (import.meta.env.VITE_IS_WEB) {
+    initWindowInfo()
+  }
 }
 // let unregistereds = createUnsubscriptionSet()
 export const initApp = () => {
@@ -31,6 +35,13 @@ export const initApp = () => {
     })
   )
   window.addEventListener('worker-initialized-main', mainWorkerResolve!)
+  appEvent.on('fullscreen', (isFullscreen) => {
+    if (isFullscreen) {
+      document.documentElement.classList.add('fullscreen')
+    } else {
+      document.documentElement.classList.remove('fullscreen')
+    }
+  })
   lyricEvent.on('titleLyricChanged', (text) => {
     if (!settingState.setting['player.isShowTitleLyric']) return
     if (text == null) {
@@ -48,6 +59,13 @@ export const initApp = () => {
         setTitle(playerState.title)
       }
     }
+    if (import.meta.env.VITE_IS_DESKTOP) {
+      // this setting only works on desktop
+      if (keys.includes('common.startInFullscreen')) {
+        setFullScreen(settings['common.startInFullscreen'] || false)
+      }
+    }
+    if (import.meta.env.VITE_IS_WEB) handleConfigChange(keys, settings)
   })
 
   document.documentElement.addEventListener(
@@ -74,11 +92,7 @@ export const initApp = () => {
   })
   if (import.meta.env.VITE_IS_WEB) {
     onRelease(() => {
-      document.body.style.removeProperty('position')
-      document.body.style.removeProperty('width')
-      document.body.style.removeProperty('height')
-      document.body.style.removeProperty('left')
-      document.body.style.removeProperty('top')
+      handleRelease()
       document.documentElement.style.fontSize = '16px'
       document.body.classList.remove('no-animation')
     })